Gate Entries behind Users

+++ app/controllers/api/v1/entries_controller.rb
@@ -1,20 +1,29 @@
 require 'pry'
 
 class Api::V1::EntriesController < ApplicationController
-  def index
-    if params[:q]
-      if !params[:q].empty?
-        query = params[:q].downcase.strip
-        keyword = Keyword.find_by(name: query)
-        entries = keyword.entries
-      else
-        entries = []
-      end
-    else
-      entries = Entry.all
-    end
+  # before_action :require_login
 
-    render json: entries, include: [:keywords]
+  def index
+    if current_user_id
+      user = User.find_by(id: current_user_id)
+
+      if user
+        render json: user.entries, status: 200
+      end
+    end
+    # if params[:q]
+    #   if !params[:q].empty?
+    #     query = params[:q].downcase.strip
+    #     keyword = Keyword.find_by(name: query)
+    #     entries = keyword.entries
+    #   else
+    #     entries = []
+    #   end
+    # else
+    #   entries = Entry.all
+    # end
+
+    # render json: entries, include: [:keywords]
   end
 
   def show
@@ -24,28 +33,47 @@ class Api::V1::EntriesController < ApplicationController
   end
 
   def create
-    entry = Entry.new(entry_params)
+    if current_user_id
+      user = User.find_by(id: current_user_id)
 
-    if entry.save
-      render json: entry, include: [:keywords]
-    else
-      render json: entry.errors.full_messages, status: :unprocessable_entity
+      if user
+        entry = user.entries.build(entry_params)
 
+        if entry.save
+          render json: entry, status: 200
+        else
+            render json: {errors: entry.errors.full_messages}, status: 400
+        end
+      end
+    end
+  end
+
+  def update
+    if current_user_id
+      entry = Entry.find_by(id: params[:id], user_id: current_user_id)
+
+      if entry && entry.update(entry_params)
+        render json: entry, status: 200
+      else
+        render json: {errors: entry.errors.full_messages}, status: 400
+      end
     end
   end
 
   def destroy
-    entry = Entry.find(params[:id])
+    if current_user_id
+      entry = Entry.find_by(id: params[:id], user_id: current_user_id)
 
-    entry.destroy
+      entry.destroy
 
-    render json: {entryId: entry.id}
+    render status: 204
+    end
   end
 
   private
 
   def entry_params
-    params.require(:entry).permit(:body, :time_interval, :keywords_attributes => :name)
+    params.require(:entry).permit(:body, :time_interval, :title, :keywords_attributes => :name)
   end
 
 end
